tests: move stepreporter tests to steplogger
Rework the StepReporter tests to use the StepLogger instead.
